Weather App/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.8
pillow==10.4.0
requests==2.32.3
ttkbootstrap==1.10.1
urllib3==2.2.2
